revised flow matrix management

+++ src/parallel_uv_method/DUAL_tree.h
@@ -3,5 +3,5 @@
 #include "../structs.h"
 #include "./parallel_structs.h"
 
-__global__ void assign_next(float * d_adjMtx_ptr, float * d_costs_ptr, 
+__global__ void assign_next(int * d_adjMtx_ptr, float * d_costs_ptr, 
                             Variable * u_vars, Variable * v_vars, int numSupplies, int numDemands);
